Refactor: extract SuccessContent from InsightsScreen
This commit extracts the `SuccessContent` composable from `InsightsScreen.kt` into its own file, `SuccessContent.kt`.

**Key Changes:**

- **`SuccessContent.kt`:**
    - Created a new internal composable function `SuccessContent` which encapsulates the UI for displaying game statistics when data is successfully loaded.
    - This includes the summary layouts (compact and expanded) and the insights table.
    - It takes `uiState`, `tableColumnsState`, `onEvent`, and `onCopySeedClick` as parameters.
    - A preview for `SuccessContent` has also been added.
- **`InsightsScreen.kt`:**
    - Removed the UI logic for the success state and replaced it with a call to the new `SuccessContent` composable.
    - The `onCopySeedClick` lambda now also includes logic to show a snackbar message "Copied seed!" when a seed is copied. This was previously handled within the table content lambda.
    - Unused imports related to the extracted UI elements were removed.

This refactoring improves the organization and readability of the `InsightsScreen` by separating the success state UI into a dedicated component.
